Create setTheRightMostUnsetBit.cpp
The function setBit(int N) is a member of the Solution class. It takes in an integer N as its input and returns the next greater integer that is a power of 2.
The code uses bitwise operators to check and set the least significant bit that is not set in N.

The first line of the function checks if N is already a power of 2. It does this by using the bitwise operator & to check if (N & (N+1)) is equal to 0.
A number is a power of 2 if and only if only one bit is set in its binary representation. The expression (N & (N+1)) will be zero if only one bit is set in N, as adding 1 to a number will toggle all the bits up to and including the rightmost set bit, resulting in a number with no set bit in common with N.

If the condition (N & (N+1)) ==0 is true, it means that N is already a power of 2 and the function returns N.
Otherwise, the function uses the bitwise operator | to set the least significant bit that is not set in N by performing a bitwise OR operation between N and (N+1). The result is returned as the next greater integer that is a power of 2.