Create swapTwoNum.cpp
This code defines a class "Solution" that contains a single function "get". The function takes in two integer arguments, "a" and "b". It then uses the bitwise XOR operator to swap the values of "a" and "b".

The bitwise XOR operator compares each bit of the first operand to the corresponding bit of the second operand. If the bits are the same, it returns 0. If they are different, it returns 1. The result is a new number where each bit is the result of the comparison.

Here, the bitwise XOR operator is used three times in succession to swap the values of "a" and "b".

On the first line, the value of "a" is assigned to "a" XOR "b".
On the second line, the value of "b" is assigned to "a" XOR "b".
On the third line, the value of "a" is assigned to "a" XOR "b".
So, after these three lines of code, the value of "a" would become the original value of "b" and the value of "b" would become the original value of "a".

Finally, the values of the swapped variables are stored in a pair, "p" and returned by the function.