Create singleNumber.cpp
This code is solving the problem of finding a single number in an array of integers where every other number appears twice, except for one. The solution first sorts the input array, then iterates through the array starting from the second element, and checks if the current element is different from the previous element. If it is, then the previous element is returned as the single number, because after sorting the array, the single number will not have a pair and will be the only element that is different from its previous element.

The time complexity of this solution is O(nlogn) because of the sorting step. Sorting algorithms such as quicksort and merge sort have an average time complexity of O(nlogn), and sorting the array is a necessary step for the solution to work correctly.

It is also worth noting that this solution would not work if all the elements in the array are the same except for one. In that case, it will return the last element, which is not the desired output. A different approach would be needed to handle this case.



USING  XOR:
This code is solving the problem of finding a single number in an array of integers where every other number appears twice, except for one. The solution uses the XOR operator to achieve this.

The XOR operator (^) returns 1 when the bits are different and 0 when they are the same. By XORing all the elements in the array together, we are effectively canceling out all the pairs of integers and the single integer will be left behind.

For example, if the array is [4, 1, 2, 1, 2], the XOR operation would be:

((4^1)^2)^(1^2) = 4

In this way, the single number 4 is left behind after XORing all the elements.

The time complexity of this solution is O(n) because it iterates through all the elements of the array once, it does not require any sorting or additional data structures, making it a more efficient solution as compared to the previous one.

Also, this solution is not dependent on the type of array, it will work for all arrays where all the elements except one are repeating twice.