Create bitDifference.cpp
The code defines a class Solution with a single function countBitsFlip. This function takes in two integers a and b, and returns the number of bits that need to be flipped to convert a into b.

The function starts by initializing a variable cnt to 0, which will be used to keep track of the number of bits that need to be flipped.

Next, the function enters a loop that iterates 32 times. This is because int values in C++ are typically represented using 32 bits.

In each iteration of the loop, the function first gets the least significant bit of a by using the bitwise AND operator & with 1. This is because 1 in binary has only the least significant bit set. Similarly, the least significant bit of b is obtained.

If the least significant bits of a and b are different, the function increments cnt to keep track of the number of bits that need to be flipped.

Finally, the function right-shifts a and b using the right-shift operator >> by 1 to get the next least significant bit. The loop continues until all 32 bits have been processed.

After the loop, the function returns the final value of cnt, which represents the number of bits that need to be flipped to convert a into b.