Create bitDiff(2ndApproach).cpp
The code defines a class Solution with two functions: countSetBit and countBitsFlip.

The countSetBit function takes in an integer n and returns the number of set bits in it. The function initializes a variable ans to 0, which will be used to keep track of the number of set bits.

Next, the function enters a loop that continues until n becomes 0. In each iteration of the loop, the function clears the least significant bit of n by using the bitwise AND operator & with (n-1). This is because n-1 will have all the bits set to the right of the least significant bit cleared and all the bits to the left of the least significant bit set. Anding n and n-1 will clear the least significant bit of n.

The function then increments ans to keep track of the number of set bits.

Finally, the function returns the final value of ans, which represents the number of set bits in n.

The countBitsFlip function takes in two integers a and b, and returns the number of bits needed to be flipped to convert a into b.

The function first calculates the XOR of a and b using the bitwise XOR operator ^. The XOR of a and b gives us a number whose set bits are the bits that are different between a and b.
Finally, the function returns the number of set bits in the XOR result by calling the countSetBit function and passing in the XOR result. This gives us the number of bits that need to be flipped to convert a into b.