Create ExceptionallyOdd.cpp
In this code, a class Solution is defined which contains a function getOddOccurrence that takes in an array arr and its length n as input. The purpose of this function is to find the number that occurs an odd number of times in the array.

The first step in the function is to create an STL map called mpp to store the frequency of each number in the array. A map is a container in the Standard Template Library (STL) in C++ that stores elements as key-value pairs. In this case, the numbers in the array will be the keys and their frequencies will be the values.

Next, the code uses a for loop to iterate through the arr array and count the frequency of each number. The mpp[arr[i]]++ line increments the frequency count of arr[i] in the map by 1.

Finally, the code uses a for-each loop to iterate through the mpp map and check the frequency of each number. The loop uses the auto keyword to automatically deduce the type of the iterator, which is a pair of int and int. The it.first and it.second variables represent the key (number) and value (frequency) respectively of each key-value pair in the map. If a number's frequency it.second is not divisible by 2 (it.second % 2 != 0), it means the frequency is odd, and the number (it.first) is returned as the result.