Create XORfromLtoR.cpp
The code defines a class Solution that contains two functions: find and findXOR.

The find function takes an integer num as input and returns an integer value by performing bitwise XOR operation on it. It uses an if-else statement to determine the value of num % 4 and returns the XOR value based on that.

The findXOR function takes two integers l and r as input and returns the XOR of all the numbers in the range [l, r]. It calls the find function to calculate the XOR value of l-1 and r, and then returns the XOR of these two values.

The XOR operation is a bitwise operator that performs the comparison of each bit of the first operand to the corresponding bit of the second operand. If both bits are different, the corresponding result bit is set to 1. If both bits are the same, the corresponding result bit is set to 0.