Create addToArrayFormIntegers.cpp
Suppose we have a vector num of integers and an integer k as shown below:
num = [2, 3, 4, 5]
k = 789
The goal of the given code is to add k to num and return the updated vector. To achieve this, the code starts from the rightmost digit of num and adds k to the corresponding digits of num, one digit at a time. If the sum is greater than or equal to 10, it sets k to the carry-over value and updates the current digit of num accordingly.

The first step is to start from the rightmost digit of num, which is 5 in this case. We add k to 5, so the sum becomes 5 + 9 = 14. Since the sum is greater than or equal to 10, we set k to the carry-over value, which is 1, and update the current digit of num to the remainder of the sum divided by 10, which is 4.

num = [2, 3, 4, 4]
k = 1

The next digit to the left is 4. We add k to 4, so the sum becomes 4 + 8 + 1 = 13. Again, the sum is greater than or equal to 10, so we set k to the carry-over value, which is 1, and update the current digit of num to the remainder of the sum divided by 10, which is 3.

num = [2, 3, 3, 4]
k = 1

We repeat this process for the next digit to the left, which is 3, and then for the leftmost digit, which is 2. At this point, we have added k to num and obtained the updated vector num = [2, 3, 3, 4]. However, there may be a carry-over value remaining, which we need to insert at the beginning of num.

In this case, the carry-over value is 1. We use a while loop to insert each digit of the carry-over value at the beginning of num, one digit at a time. After the loop, we obtain the final vector num = [1, 2, 3, 3, 4].

Therefore, the output of the given code for the input num = [2, 3, 4, 5] and k = 789 is [1, 2, 3, 3, 4].