Create PrimeFactors.cpp
The AllPrimeFactors function takes an integer N as input and returns a vector of all its prime factors. The algorithm first initializes an empty vector ans to store the prime factors and an unordered set s to keep track of which factors have already been added to the vector. It then loops through all possible prime factors less than or equal to the square root of N.

While a prime factor i is a factor of N, it is added to the ans vector and inserted into the set s if it hasn't already been added. The N value is divided by i so that we can continue finding the remaining prime factors. This continues until there are no more factors of i in N.

After the loop finishes, if there is still a factor greater than the square root of N, it is added to the ans vector. Finally, the function returns the vector of prime factors.