Create AllDivisorsOfNum.cpp
This code defines a class called Solution with a single method called print_divisors. The print_divisors method takes an integer n as input and prints all of the divisors of n to the console in ascending order.

To find the divisors of n, the method creates an empty set of integers called s and loops through all integers from 1 to the square root of n. For each integer i, it checks if i is a divisor of n by checking if n % i == 0. If i is a divisor of n, the method adds both i and n/i to the set s.

Finally, the method loops through the set s and prints each element to the console, separated by a space.