Create checkPrime.cpp
This code implements the Sieve of Eratosthenes algorithm to generate all prime numbers between 2 and a given integer N.

The Sieve of Eratosthenes algorithm works by starting with a list of all integers between 2 and N, and successively crossing out multiples of each prime number found. The numbers that remain in the list are all prime.

In this implementation, the algorithm is split into two functions:

The isPrime function checks whether a given number is prime or not. It does this by iterating over all possible divisors of the number, from 2 up to the square root of the number. If any divisor is found, other than 1 or the number itself, then the number is not prime and the function returns false. If no divisors are found, the number is prime and the function returns true.

The sieveOfEratosthenes function generates a vector of all prime numbers between 2 and N. It starts by initializing an empty vector to store the prime numbers. It then iterates over all numbers from 2 up to N, checking each number for primality using the isPrime function. If a number is found to be prime, it is added to the vector. The resulting vector contains all prime numbers between 2 and N.

The implementation is straightforward and easy to understand, but not very efficient for very large values of N. The algorithm can be optimized by using a boolean array to keep track of which numbers have been crossed out, rather than iterating over all possible divisors.