Create checkIfTheArrayIsSorted.cpp
This code defines a class Solution with a single public method check that takes a vector of integers as input and returns a boolean value. The goal of the method is to determine whether the input vector can be sorted into non-decreasing order by at most one swap of two adjacent elements.

To accomplish this, the method first creates a new vector vec with the same size as the input vector. It then initializes a counter cnt to zero and iterates through the input vector using a for loop.

For each element in the input vector, the method compares it to the next element (wrapping around to the beginning of the vector if necessary) and checks whether it is greater than the next element. If so, it increments the cnt counter to indicate that there is a pair of out-of-order elements.

Finally, the method returns true if the cnt counter is less than or equal to 1 (meaning that at most one swap is needed to put the elements in non-decreasing order) and false otherwise (meaning that more than one swap is needed, or that the input vector is already in non-decreasing order).