Create BestTimeToBuyAndSellStock.cpp
The algorithm used in the code is as follows:

Initialize the variables max_profit, low, and curr to 0 and INT_MAX, respectively. max_profit will store the maximum profit that can be made, low will store the lowest price seen so far, and curr will store the current profit.

Loop through the prices vector. For each price prices[i]:

a. If prices[i] is lower than the lowest price seen so far (low), update low to prices[i].

b. Calculate the maximum profit that can be made with the current price prices[i] as the selling price and low as the buying price, by subtracting low from prices[i]. Store this value in max_profit.

c. If max_profit is greater than curr, update curr to max_profit.

Return curr, which stores the maximum profit that can be made.

The time complexity of the algorithm is O(n), where n is the number of prices in the prices vector, as we loop through the prices vector once. The space complexity of the algorithm is O(1), as we use only constant space to store the variables.