Create MaxConsecutiveOnes.cpp
This code takes in a vector of integers (nums) and returns the maximum count of consecutive ones in the vector. It does so by iterating through the vector, keeping track of the count of consecutive ones (cnt) and updating the maximum count (maxi) as it goes along.

Each time it encounters a 1, it increments the count of consecutive ones. If it encounters a 0, it resets the count of consecutive ones to 0. Finally, it returns the maximum count of consecutive ones.