Create kokoEatingBananas.cpp
Koko loves to eat bananas. There are n piles of bananas, the ith pile has piles[i] bananas. The guards have gone and will come back in h hours. Koko can decide her bananas-per-hour eating speed of k. Each hour, she chooses some pile of bananas and eats k bananas from that pile. If the pile has less than k bananas, she eats all of them instead and will not eat any more bananas during this hour. Koko likes to eat slowly but still wants to finish eating all the bananas before the guards come back. Return the minimum integer value of k such that she can eat all the bananas within h hours.

The given code uses a binary search approach to find the minimum value of k such that Koko can eat all the bananas within h hours. The key idea is to use binary search on the range of possible values of k, and for each possible value of k, check if Koko can eat all the bananas within h hours.

Here's how the code works:

The function minEatingSpeed takes the input vector piles and integer h as arguments and returns the minimum integer value of k such that Koko can eat all the bananas within h hours.

In this function, we initialize two integer variables left and right. We use left to store the minimum possible value of k, which is 1, and right to store the maximum possible value of k, which is the maximum value of piles. Here, 1e9 is used as an upper bound on right since this is the maximum number of bananas in any pile.

We use a while loop to perform binary search on the range [left, right] until only one possible value of k remains.

In each iteration of the while loop, we calculate the midpoint of the range [left, right] using the formula mid = left + (right - left) / 2.

We then call the function canEatAll with the current value of k = mid. If canEatAll returns true, we update the upper bound of the range to mid, since we want to find the minimum value of k that satisfies the condition. Otherwise, we update the lower bound of the range to mid + 1.

After the while loop terminates, we return the value of left, which is the minimum integer value of k that satisfies the condition.

The function canEatAll takes the input vector piles, integer h, and integer k as arguments and returns a boolean value. This function checks if Koko can eat all the bananas within h hours at a given value of k.

We initialize an integer variable hours to 0, which we will use to keep track of the total number of hours needed to eat all the bananas.

We use a for loop to iterate over each pile of bananas. For each pile, we calculate the number of hours needed to eat that pile at the given speed k using the formula (pile + k - 1) / k, which rounds up the division to the nearest integer. We add this value to the hours variable.

Finally, we check if the total number of hours needed to eat all the bananas is less than the targeted hours.