Create numOfZeroFilledSubarray.cpp
The zeroFilledSubarray function takes in a vector nums of integers, and calculates the number of subarrays in which all the elements are 0. It does this by iterating through the vector, keeping track of the number of consecutive 0s using the variable cnt, and adding the number of subarrays with 0s to the variable ans when a non-zero element is encountered.

Finally, the function adds the number of subarrays with 0s at the end of the vector (if any) to ans, and returns the total number of subarrays with 0s.