Create numOfSubstringsWithOnly1.cpp
This code defines a function called numSub that takes a string s as input and returns the number of substrings of s that contain only '1's. It does this by looping through the string, keeping track of the length of consecutive '1's using the variable cnt, and adding the number of possible substrings to ans whenever it encounters a '0'.

The number of possible substrings can be calculated using the formula (n * (n+1))/2, where n is the length of the consecutive '1's. This formula is used to count the number of substrings that can be formed by choosing any two indices i and j where i <= j, in a sequence of length n.

Finally, the modulo operator % is used to ensure that the result is within the range of a 32-bit signed integer, and the result is returned.