Create ConsecutiveCharacters.cpp
This code is implementing a function called maxPower that takes a string s as input and returns the maximum count of consecutive characters in the string.

The function first initializes two variables maxi and cnt to store the maximum count of consecutive characters seen so far and the current count of consecutive characters, respectively. maxi is initialized to the smallest possible integer value, since any count of consecutive characters seen so far will be greater than this value.

The function then iterates through each character in the string s. If the current character is the same as the next character, the function increments the count of consecutive characters (cnt). If the current character is different from the next character, the function updates the maximum count of consecutive characters seen so far (maxi) if the current count (cnt) is greater, and resets the count of consecutive characters to 1.

Finally, the function returns the maximum count of consecutive characters (maxi) seen in the string s.

For example, if the input string is "aabbbbccccddddd", the function will return 5, since the maximum count of consecutive characters is 5 (there are 5 consecutive 'd' characters).