Create maximumSubarray.cpp
The Kadane's algorithm is based on the observation that, for any given index i in the array, the maximum subarray ending at that index either includes the element at index i, or it doesn't.

If it doesn't include the element at index i, then the maximum subarray ending at index i is simply the maximum subarray ending at the previous index i-1. However, if it does include the element at index i, then the maximum subarray ending at index i is the maximum of two possibilities: either the element at index i itself forms the maximum subarray, or it is part of a maximum subarray that ends at a previous index.

Thus, to find the maximum subarray sum, we simply need to keep track of the maximum subarray sum ending at each index i in the array, and take the maximum of all these values as our final result.

To achieve this, we use two variables: maxi and sum. The variable maxi keeps track of the maximum subarray sum seen so far, while the variable sum keeps track of the current sum. We initialize maxi to be the minimum possible integer value (INT_MIN), so that it will be updated by the first element of the array.

We then loop through each element of the input array using a for loop. For each element, we add it to the running sum sum. We then update maxi to be the maximum of sum and maxi. If sum becomes negative at any point, we reset it to zero, because adding a negative number to the sum will only decrease it further.

At the end of the loop, maxi contains the maximum subarray sum seen in the entire array, which we return as our final result.

As for time complexity, the algorithm uses a single loop over the entire input array, making it a linear time algorithm with a time complexity of O(n), where n is the length of the input array. This makes it a very efficient solution for finding the maximum subarray sum.