Create AlternateNumbers.cpp
This code takes an input vector a of integers and rearranges it so that positive and negative integers alternate with each other in the output vector. If there are more positive integers than negative integers, any extra positive integers are appended to the end of the output vector. If there are more negative integers than positive integers, any extra negative integers are appended to the end of the output vector.

The code first creates two empty vectors pos and neg to hold positive and negative integers respectively. It then iterates over each integer in the input vector and adds it to the appropriate vector. Positive integers are added to the pos vector and negative integers are added to the neg vector.

After the positive and negative integers have been separated into two vectors, the code checks if there are more positive integers than negative integers. If there are, it iterates over the negative integers and alternates them with the positive integers in the output vector. It does this by placing the positive integer at even indices and the negative integer at odd indices. If there are any remaining positive integers, it appends them to the end of the output vector.

If there are more negative integers than positive integers, the code follows a similar procedure but iterates over the positive integers instead. It alternates the positive integers with the negative integers in the output vector and appends any remaining negative integers to the end of the output vector.

Finally, the code returns the modified input vector.