Create SuperiorElements.cpp
The given code solves the problem of finding superior elements in a vector. An element in the vector is superior if it is greater than all the elements to its right in the vector.

To solve the problem, the function superiorElements() takes a vector of integers as input. It first initializes an empty vector called ans which will store the superior elements. It also initializes a variable called maxi to the smallest possible integer value.

Then, it traverses the input vector from right to left using a for loop. For each element, it checks whether it is greater than the maximum value maxi found so far. If it is, then it adds the element to the ans vector. Then, it updates the maxi variable to the maximum value seen so far.

Finally, it sorts the ans vector in ascending order using the sort() function and returns it.

The time complexity of this function is O(nlogn), where n is the size of the input vector, due to the sorting operation. However, the space complexity is O(m), where m is the number of superior elements found in the vector, as only those elements are stored in the ans vector.