Create leadersInArray.cpp
Leaders are defined as the elements in the array that are greater than or equal to all the elements to their right.

The function leaders takes an array a[] of size n as input and returns a vector of integers ans that contains all the leaders in the array.

The first approach that is commented out in the code, iterates through the array from left to right and for each element, checks if there are any greater elements to its right. If not, the element is a leader and is added to the ans vector. However, this approach has a time complexity of O(n^2), which is not efficient for large arrays.

The second approach, which is implemented, iterates through the array from right to left and maintains a variable maxi that keeps track of the maximum element encountered so far. If an element is greater than or equal to maxi, it is a leader and is added to the ans vector. This approach has a time complexity of O(n), which is more efficient than the first approach.

Finally, the ans vector is returned after reversing it, since the leaders are added in reverse order while iterating the array from right to left.